fix bug: email already verified and no need to send otp email

+++ src/controllers/auth.ts
@@ -71,6 +71,9 @@ export const sendReVerificationToken: RequestHandler = async (req, res) => {
   const user = await User.findById(userId);
   if (!user) return res.status(403).json({ error: 'Invalid request!' });
 
+  if (user.verified)
+    return res.status(422).json({ error: 'Email already verified!' });
+
   await EmailVerificationToken.findOneAndDelete({ owner: userId });
 
   const token = generateToken();
